+++ src/access_log.c
@@ -245,8 +245,8 @@ accessLogSquid(AccessLogEntry * al)
 	client = inet_ntoa(al->cache.caddr);
     user = accessLogFormatName(al->cache.authuser ?
 	al->cache.authuser : al->cache.rfc931);
-    logfilePrintf(logfile, "%9d.%03d %6d %s %s/%03d %lu %s %s %s %s%s/%s %s",
-	(int) current_time.tv_sec,
+    logfilePrintf(logfile, "%9ld.%03d %6d %s %s/%03d %lu %s %s %s %s%s/%s %s",
+	(long int) current_time.tv_sec,
 	(int) current_time.tv_usec / 1000,
 	al->cache.msec,
 	client,
